requirements.txt
streamlit>=0.82.0
streamlit-webrtc
numpy>=1.19.5
pandas>=1.2.4
speechrecognition>=3.8.1
streamlit-audiorecorder
pydub
SpeechRecognition
